requirements.txt
azure-functions
pytestcd 